logica finish round

+++ src/main/java/com/reccos/admin/model/League.java
@@ -57,7 +57,7 @@ public class League {
 	private Integer qt_group;
 	
 	@Column(name = "round_actv")
-	private Integer round_actv;
+	private long round_actv;
 
 	@Column(name = "status")
 	private Boolean status;
@@ -82,8 +82,8 @@ public class League {
 	}
 
 	public League(long id, String name, LocalDate dt_start, LocalDate dt_end, long idd_fed, String league_system,
-			String league_mode, Integer max_teams, Integer min_teams, Integer qt_group, Integer round_actv,
-			Boolean status, Set<Group> groups, Set<Team> teams, Federation federation) {
+			String league_mode, Integer max_teams, Integer min_teams, Integer qt_group, long round_actv, Boolean status,
+			Set<Group> groups, Set<Team> teams, Federation federation) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -214,11 +214,11 @@ public class League {
 		this.groups = groups;
 	}
 
-	public Integer getRound_actv() {
+	public long getRound_actv() {
 		return round_actv;
 	}
 
-	public void setRound_actv(Integer round_actv) {
+	public void setRound_actv(long round_actv) {
 		this.round_actv = round_actv;
 	}
